I FINALLY GOT THE TEXTURES TO RENDER INTO TEXTURE2D OBJECTS CORRECTLY!
The code in Patch.cs is still a mess, but I left the old code in just to have a copy in the repo before I delete it all.
Next up, I need to fix an apparent issue with the UV generation code, as some faces do not have proper texture alignment now.